break-out TypeFilterGroup and TypeFilterRow
